Add condition to IAM Policy to match Role Condition
The user is able to pass in a reference to an existing IAM Role as an
alternative to creating one; a Policy was always created and attached to
the role whether created or supplied. With this change, the passed role
is assumed to have requisite policies already attached so that no IAM
resources are created if the Role is passed by reference.

Additionally, the InstanceProfile is tagged with the /account-managed/
path if created alongside a pre-existing Role to facilitate tracking